requirements.txt
fastapi[standard]
jinja2
sqlmodel
argon2-cffi
PyMySQL
cryptography